python/01-tutorials/01-agentkit-runtime/callback/pyproject.toml
[project]
name = "callback-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "agentkit-sdk-python==0.4.3",
  "google-adk==1.21.0",
  "veadk-python==0.5.20",
]
